Made an artist/painting card
This card has both a light and dark theme option. Added the necessary viewports to the preview.ts file of the .storybook folder and connected the general styles of the project. In the main.ts file of the same folder,  added the svg loading function. Added missing rule to .eslintrc