Code refactoring
Components text button and text button with icon combined into one. Their styles are customizable with props. Removed the text button component with an icon as unnecessary.

Removed random space in MainPageCard component styles. The interface of the MainPageCard component was moved to a separate file. Removed the static path to the image from the img tag, now it is taken from the props.

For the Link component, removed the underlining of the text if an image is passed to it

In the styles of the GoUpButton and IconButton components, made the border-radius always round, and also grouped the styles for :hover and :focus-visible

Added a logo

The ChangeThemeButton component now accepts classes from outside for editing